refactor: move ticket link component
