+++ app/src/features/ui/index.ts
@@ -4,3 +4,4 @@ export * from './InfoBubble'
 export * from './AvatarDiv'
 export * from './Error404'
 export * from './ExternalLink'
+export * from './TicketLink'
