+++ app/src/pages/ImprovementPage.tsx
@@ -5,9 +5,8 @@ import { IMPROVEMENT_STATUS_UI } from '../features/items'
 import { renderDate } from '../data/util'
 import { ImprovementIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
-import { Error404 } from '../features/ui'
+import { Error404, TicketLink } from '../features/ui'
 import { useParams } from 'react-router-dom'
-import { TicketLink } from '../components/TicketLink'
 import { ImprovementId } from '../data/types'
 import { ItemDetailsPage, useEditorStore } from '../features/items'
 import { Markdown } from '../features/markdown'
